in progress

+++ src/formoptions.cc
@@ -3,8 +3,8 @@
 // e-mail:   Jean-Charles.Lambert@lam.fr                                      
 // address:  Centre de donneeS Astrophysique de Marseille (CeSAM)              
 //           Laboratoire d'Astrophysique de Marseille                          
-//           Pôle de l'Etoile, site de Château-Gombert                         
-//           38, rue Frédéric Joliot-Curie                                     
+//           Pï¿½le de l'Etoile, site de Chï¿½teau-Gombert                         
+//           38, rue Frï¿½dï¿½ric Joliot-Curie                                     
 //           13388 Marseille cedex 13 France                                   
 //           CNRS U.M.R 7326                                                   
 // ============================================================================
@@ -257,7 +257,7 @@ void FormOptions::update()
 // updateFrame                                                                 
 void FormOptions::updateFrame(const int frame, const int tot)
 {
-  form.fps_lcd->display(1000*frame/time.elapsed());
+  form.fps_lcd->display((int) (1000*frame/time.elapsed()));
   form.total_lcd->display(tot);
   time.restart();
 }
